fix: make NNs pickleable
