+++ src/safeds/ml/nn/layers/_lstm_layer.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import sys
 from typing import TYPE_CHECKING, Any
 
-from safeds._config import _init_default_device
 from safeds._utils import _structural_hash
 from safeds._validation import _check_bounds, _ClosedBound
 from safeds.ml.nn.typing import ModelImageSize
@@ -11,7 +10,7 @@ from safeds.ml.nn.typing import ModelImageSize
 from ._layer import Layer
 
 if TYPE_CHECKING:
-    from torch import Tensor, nn
+    from torch import nn
 
 
 class LSTMLayer(Layer):
@@ -37,6 +36,8 @@ class LSTMLayer(Layer):
         self._output_size = neuron_count
 
     def _get_internal_layer(self, **kwargs: Any) -> nn.Module:
+        from ._internal_layers import _InternalLSTMLayer  # Slow import on global level
+
         if "activation_function" not in kwargs:
             raise ValueError(
                 "The activation_function is not set. The internal layer can only be created when the activation_function is provided in the kwargs.",
@@ -47,7 +48,7 @@ class LSTMLayer(Layer):
         if self._input_size is None:
             raise ValueError("The input_size is not yet set.")
 
-        return _create_internal_model(self._input_size, self._output_size, activation_function)
+        return _InternalLSTMLayer(self._input_size, self._output_size, activation_function)
 
     @property
     def input_size(self) -> int:
@@ -94,30 +95,3 @@ class LSTMLayer(Layer):
 
     def __sizeof__(self) -> int:
         return sys.getsizeof(self._input_size) + sys.getsizeof(self._output_size)
-
-
-def _create_internal_model(input_size: int, output_size: int, activation_function: str) -> nn.Module:
-    from torch import nn
-
-    _init_default_device()
-
-    class _InternalLayer(nn.Module):
-        def __init__(self, input_size: int, output_size: int, activation_function: str):
-            super().__init__()
-            self._layer = nn.LSTM(input_size, output_size)
-            match activation_function:
-                case "sigmoid":
-                    self._fn = nn.Sigmoid()
-                case "relu":
-                    self._fn = nn.ReLU()
-                case "softmax":
-                    self._fn = nn.Softmax()
-                case "none":
-                    self._fn = None
-                case _:
-                    raise ValueError("Unknown Activation Function: " + activation_function)
-
-        def forward(self, x: Tensor) -> Tensor:
-            return self._fn(self._layer(x)[0]) if self._fn is not None else self._layer(x)[0]
-
-    return _InternalLayer(input_size, output_size, activation_function)
